Reorganize spec_helper and introduce modern rspec niceties
