+++ .gitignore
@@ -53,3 +53,5 @@ deploy/.rvmrc
 config/settings.local.yml
 config/settings/*.local.yml
 config/environments/*.local.yml
+
+spec/examples.txt
