+++ spec/support/dor_config.rb
@@ -1,0 +1,1 @@
+Dor::Config.configure.suri.mint_ids = false
